chore!: Drop support for all legacy protocols
All network protocols will now run exclusively via the new network
stack over libp2p.

BREAKING CHANGE.
