+++ daemon/src/order.rs
@@ -1,5 +1,6 @@
 pub const PROTOCOL: &str = "/itchysats/order/1.0.0";
 
+mod contract_setup;
 pub mod maker;
 mod protocol;
 pub mod taker;
