+++ daemon/src/order/taker.rs
@@ -1,15 +1,15 @@
 use crate::command;
 use crate::oracle;
 use crate::oracle::NoAnnouncement;
+use crate::order::contract_setup;
 use crate::order::protocol::Decision;
 use crate::order::protocol::MakerMessage;
+use crate::order::protocol::SetupMsg;
 use crate::order::protocol::TakerMessage;
 use crate::order::PROTOCOL;
 use crate::process_manager;
 use crate::projection;
-use crate::setup_contract;
 use crate::wallet;
-use crate::wire::SetupMsg;
 use anyhow::bail;
 use anyhow::Context;
 use anyhow::Result;
@@ -174,7 +174,7 @@ impl Actor {
                     .send(oracle::GetAnnouncements(vec![oracle_event_id]))
                     .await??;
 
-                let dlc = setup_contract::new(
+                let dlc = contract_setup::new(
                     sink.with(|msg| future::ok(TakerMessage::ContractSetupMsg(Box::new(msg)))),
                     Box::pin(stream.filter_map(|msg| async move {
                         let msg = match msg {
