+++ maker-frontend/src/components/Types.tsx
@@ -97,10 +97,6 @@ export class State {
                 return "Close Proposed";
             case StateKey.OUTGOING_SETTLEMENT_PROPOSAL:
                 return "Close Proposed";
-            case StateKey.INCOMING_ROLLOVER_PROPOSAL:
-                return "Rollover Proposed";
-            case StateKey.OUTGOING_ROLLOVER_PROPOSAL:
-                return "Rollover Proposed";
             case StateKey.ROLLOVER_SETUP:
                 return "Rollover Setup";
             case StateKey.PENDING_REFUND:
@@ -141,8 +137,6 @@ export class State {
             case StateKey.CONTRACT_SETUP:
             case StateKey.OUTGOING_SETTLEMENT_PROPOSAL:
             case StateKey.INCOMING_SETTLEMENT_PROPOSAL:
-            case StateKey.INCOMING_ROLLOVER_PROPOSAL:
-            case StateKey.OUTGOING_ROLLOVER_PROPOSAL:
             case StateKey.ROLLOVER_SETUP:
             case StateKey.PENDING_OPEN:
             case StateKey.REFUNDED:
@@ -166,7 +160,6 @@ export class State {
             case StateKey.OPEN_COMMITTED:
             case StateKey.PENDING_REFUND:
             case StateKey.OUTGOING_SETTLEMENT_PROPOSAL:
-            case StateKey.OUTGOING_ROLLOVER_PROPOSAL:
             case StateKey.PENDING_CET:
             case StateKey.PENDING_CLOSE:
                 return StateGroupKey.OPEN;
@@ -174,7 +167,6 @@ export class State {
             case StateKey.INCOMING_SETTLEMENT_PROPOSAL:
                 return StateGroupKey.PENDING_SETTLEMENT;
 
-            case StateKey.INCOMING_ROLLOVER_PROPOSAL:
             case StateKey.ROLLOVER_SETUP:
                 return StateGroupKey.PENDING_ROLLOVER;
 
@@ -195,8 +187,6 @@ export enum Action {
     ROLL_OVER = "rollOver",
     ACCEPT_SETTLEMENT = "acceptSettlement",
     REJECT_SETTLEMENT = "rejectSettlement",
-    ACCEPT_ROLLOVER = "acceptRollover",
-    REJECT_ROLLOVER = "rejectRollover",
 }
 
 const enum StateKey {
@@ -211,8 +201,6 @@ const enum StateKey {
     OPEN_COMMITTED = "OpenCommitted",
     OUTGOING_SETTLEMENT_PROPOSAL = "OutgoingSettlementProposal",
     INCOMING_SETTLEMENT_PROPOSAL = "IncomingSettlementProposal",
-    OUTGOING_ROLLOVER_PROPOSAL = "OutgoingRolloverProposal",
-    INCOMING_ROLLOVER_PROPOSAL = "IncomingRolloverProposal",
     ROLLOVER_SETUP = "RolloverSetup",
     PENDING_REFUND = "PendingRefund",
     REFUNDED = "Refunded",
