+++ maker/src/lib.rs
@@ -11,10 +11,6 @@ pub use actor_system::ActorSystem;
 
 mod actor_system;
 pub mod cfd;
-mod collab_settlement;
-mod connection;
-mod contract_setup;
-mod legacy_rollover;
 mod metrics;
 pub mod routes;
